consolidate input batch
